Refactoring: rename a few things
