+++ config/settings/base.py
@@ -82,7 +82,7 @@ LOCAL_APPS = [
     "envergo.confs.apps.ConfsConfig",
     "envergo.admin.config.EnvergoAdminConfig",
     "envergo.urlmappings",
-    "envergo.haies",
+    "envergo.hedges",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
