+++ envergo/hedges/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class HaiesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "envergo.haies"
+    name = "envergo.hedges"
