+++ envergo/hedges/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 
-from envergo.haies.views import Saisie, SaveHedgeDataView
+from envergo.hedges.views import HedgeInput, SaveHedgeDataView
 
 urlpatterns = [
-    path("saisie/", Saisie.as_view(), name="input_hedges"),
+    path("input/", HedgeInput.as_view(), name="input_hedges"),
     path("save/", SaveHedgeDataView.as_view(), name="save_hedges"),
 ]
