+++ envergo/templates/hedges/input.html
@@ -148,7 +148,7 @@
   <script defer src="{% static 'leaflet-editable/src/Leaflet.Editable.js' %}"></script>
   <script defer src="{% static 'js/libs/leaflet-icon-fix.js' %}"></script>
   <script defer src="{% static 'vue/dist/vue.global.js' %}"></script>
-  <script type="module" src="{% static 'saisie/app.js' %}"></script>
+  <script type="module" src="{% static 'input/app.js' %}"></script>
 {% endblock %}
 
 {% block css %}
